Extract tokenize helpers and move wrap tests
